Create source troves from scratch to avoid copying undesireable troveinfo.
This fixes a bug where versions would be rewritten back in time during a
promote, because every version of a source mangled from an upstream source
with a non-empty clonedfrom entry would also carry that clonedfrom entry even
though the contents were not identical. This caused the promote to try to
reuse the version number of the first promote even when the version numbers
were completely different.

See also CNY-3181